requirements.txt
streamlit>=1.28.0
requests>=2.31.0
supabase>=2.0.0
pyyaml>=6.0
